% Auto-generated: do not edit by hand
\name{moleculeItem}

\alias{moleculeItem}

\title{MoleculeItemReact component}

\description{

}

\usage{
moleculeItem(value=NULL)
}

\arguments{
\item{value}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

